move lifecycle and scope back into html as per review
