+++ packages/yew/src/dom_bundle/mod.rs
@@ -17,7 +17,7 @@ mod btext;
 #[cfg(test)]
 mod tests;
 
-use self::bcomp::{BComp, Scoped};
+use self::bcomp::BComp;
 use self::blist::BList;
 use self::bnode::BNode;
 use self::bportal::BPortal;
@@ -25,19 +25,18 @@ use self::bsuspense::BSuspense;
 use self::btag::BTag;
 use self::btext::BText;
 
-pub(crate) use self::bcomp::{Mountable, PropsWrapper};
+pub(crate) use self::bcomp::{ComponentRenderState, Mountable, PropsWrapper, Scoped};
 pub(crate) use self::btag::{InputFields, Value};
 
 #[doc(hidden)] // Publically exported from crate::app_handle
 pub use self::app_handle::AppHandle;
-#[doc(hidden)] // Publically exported from crate::html
-pub use self::bcomp::{AnyScope, Scope, SendAsMessage};
 #[doc(hidden)] // Publically exported from crate::events
 pub use self::btag::set_event_bubbling;
 #[cfg(test)]
 #[doc(hidden)] // Publically exported from crate::tests
 pub use self::tests::layout_tests;
 
+use crate::html::AnyScope;
 use crate::NodeRef;
 use web_sys::{Element, Node};
